+++ commands.js
@@ -8,6 +8,7 @@ import { resetFishingProfitTracker } from "./features/overlays/fishingProfitTrac
 import { resetDropNumbers } from "./features/chat/messageOnDrop";
 import { calculateFishingPetPrices } from "./features/commands/calculateFishingPetsPrices";
 import { calculateGearCraftPrices } from "./features/commands/calculateGearCraftPrices";
+import { moveAllGuis } from "./moveAllOverlays";
 
 register("command", (...args) => {
     settings.openGUI();
@@ -59,3 +60,7 @@ register("command", (...args) => {
 register("command", (...args) => {
     calculateGearCraftPrices();
 }).setName("feeshGearCraftPrices");
+
+register("command", (...args) => {
+    moveAllGuis();
+}).setName("feeshMoveAllGuis");